cellular automata playback

+++ core/src/com/pixelatedmind/game/tinyuniverse/generation/music/song/model/SongModel.kt
@@ -1,7 +1,7 @@
 package com.pixelatedmind.game.tinyuniverse.generation.music.song.model
 
 class SongModel(val name : String,
-                val beatsPerMinute : Float,
+                val beatsPerMinute : Int,
                 val timeSignatureBottomNumber : Int,
                 val instrumentStreams : List<InstrumentStreamModel>) {
